requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Werkzeug==3.0.1
Flask-WTF==1.2.1
WTForms==3.1.1
email-validator==2.1.0.post1
python-dotenv==1.0.0
Pillow==10.1.0
faiss-cpu==1.11.0
numpy==1.26.4
pdfplumber==0.11.2
sentence-transformers==3.0.1
python-docx==1.1.2
pandas==2.2.2
openai==1.40.0
requests==2.32.3
pydantic==2.11.7
langchain==0.3.26
Flask
gunicorn
# if you use Postgres:
psycopg2-binary
SQLAlchemy
alembic
httpx
python-dotenv
beautifulsoup4==4.12.3
